Created login/signup components and organized them
