+++ capstone-ui/src/main.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './App.jsx'
+import App from "./Components/App/App"
 import './index.css'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
